feat: `cargo e -s i` install . or explicit path; this filters the output of cargo install.
- tested spawned filtered install on ubuntu 24.10; macos 15.2; windows 10 / 11.
- powershell to spawn a separate install to allow for cargo-e to install
  itself.  this may work on other platforms with powershell
  installed.  Other platforms linux/macos may not require the
  acrobatics.
- add stdout and stderr argument processing for improved diagnostics;
  the event dispatchers are now operating on file input instead of just
  realtime cargo output processing.
- CLI struct in e_cli.rs to include options for stdout and stderr capture paths
- Refactored command handling in main.rs to support diagnostics processing after install commands
- e_diagnostics_dispatchers.rs to break out handling diagnostic callbacks and message processing

cargo has cargo install.  There is no short for install,  no
short i.  With cargo-e; there is -s to specify the subcommand.  `-s i` or
`--subcommand install` proceeds just like `cargo install --path .`  It
gets filtered; errors/warning are numbered;

errors appear at the bottom.

If you specify an explicit name, that name will be used as a path to the
path argument.

the install is detached.  the filtered output will flow to the terminal when complete.